[CHORE] Move travel into its own directory
